Simplified the model_creation module by splitting into files
