+++ binding/python3/model_creation/axis.py
@@ -1,0 +1,33 @@
+from enum import Enum
+
+import numpy as np
+
+from .marker import Marker
+
+
+class Axis:
+    class Name(Enum):
+        X = 0
+        Y = 1
+        Z = 2
+
+    def __init__(self, name: Name, start: Marker, end: Marker):
+        """
+        Parameters
+        ----------
+        name:
+            The AxisName of the Axis
+        start:
+            The initial Marker
+        """
+        self.name = name
+        self.start_point = start
+        self.end_point = end
+
+    def axis(self) -> np.ndarray:
+        """
+        Returns the axis vector
+        """
+        start = self.start_point.position
+        end = self.end_point.position
+        return end - start
